Resolve circular dependencies on JS side (#2970)
This PR removes circular dependencies and adds a workflow which checks for new ones.
