+++ .github/workflows/static-root-checks.yml
@@ -29,3 +29,5 @@ jobs:
         run: yarn tsc --noEmit
       - name: Lint
         run: yarn lint:js-root
+      - name: Check for circular dependencies
+        run: yarn circularDependencyCheck
