+++ src/handlers/RotationGestureHandler.ts
@@ -1,35 +1,9 @@
+import { RotationGestureHandlerEventPayload } from './GestureHandlerEventPayload';
 import createHandler from './createHandler';
 import {
   BaseGestureHandlerProps,
   baseGestureHandlerProps,
 } from './gestureHandlerCommon';
-
-export type RotationGestureHandlerEventPayload = {
-  /**
-   * Amount rotated, expressed in radians, from the gesture's focal point
-   * (anchor).
-   */
-  rotation: number;
-
-  /**
-   * X coordinate, expressed in points, of the gesture's central focal point
-   * (anchor).
-   */
-  anchorX: number;
-
-  /**
-   * Y coordinate, expressed in points, of the gesture's central focal point
-   * (anchor).
-   */
-  anchorY: number;
-
-  /**
-   *
-   * Instantaneous velocity, expressed in point units per second, of the
-   * gesture.
-   */
-  velocity: number;
-};
 
 export interface RotationGestureHandlerProps
   extends BaseGestureHandlerProps<RotationGestureHandlerEventPayload> {}
