+++ src/handlers/getNextHandlerTag.ts
@@ -1,0 +1,5 @@
+let handlerTag = 1;
+
+export function getNextHandlerTag(): number {
+  return handlerTag++;
+}
