+++ src/index.ts
@@ -23,35 +23,24 @@ export { MouseButton } from './handlers/gestureHandlerCommon';
 export type { GestureType } from './handlers/gestures/gesture';
 export type {
   TapGestureHandlerEventPayload,
-  TapGestureHandlerProps,
-} from './handlers/TapGestureHandler';
-export type {
   ForceTouchGestureHandlerEventPayload,
-  ForceTouchGestureHandlerProps,
-} from './handlers/ForceTouchGestureHandler';
-export type { ForceTouchGestureChangeEventPayload } from './handlers/gestures/forceTouchGesture';
-export type {
   LongPressGestureHandlerEventPayload,
-  LongPressGestureHandlerProps,
-} from './handlers/LongPressGestureHandler';
-export type {
   PanGestureHandlerEventPayload,
-  PanGestureHandlerProps,
-} from './handlers/PanGestureHandler';
-export type { PanGestureChangeEventPayload } from './handlers/gestures/panGesture';
-export type {
   PinchGestureHandlerEventPayload,
-  PinchGestureHandlerProps,
-} from './handlers/PinchGestureHandler';
-export type { PinchGestureChangeEventPayload } from './handlers/gestures/pinchGesture';
-export type {
   RotationGestureHandlerEventPayload,
-  RotationGestureHandlerProps,
-} from './handlers/RotationGestureHandler';
-export type {
+  NativeViewGestureHandlerPayload,
   FlingGestureHandlerEventPayload,
-  FlingGestureHandlerProps,
-} from './handlers/FlingGestureHandler';
+} from './handlers/GestureHandlerEventPayload';
+export type { TapGestureHandlerProps } from './handlers/TapGestureHandler';
+export type { ForceTouchGestureHandlerProps } from './handlers/ForceTouchGestureHandler';
+export type { ForceTouchGestureChangeEventPayload } from './handlers/gestures/forceTouchGesture';
+export type { LongPressGestureHandlerProps } from './handlers/LongPressGestureHandler';
+export type { PanGestureHandlerProps } from './handlers/PanGestureHandler';
+export type { PanGestureChangeEventPayload } from './handlers/gestures/panGesture';
+export type { PinchGestureHandlerProps } from './handlers/PinchGestureHandler';
+export type { PinchGestureChangeEventPayload } from './handlers/gestures/pinchGesture';
+export type { RotationGestureHandlerProps } from './handlers/RotationGestureHandler';
+export type { FlingGestureHandlerProps } from './handlers/FlingGestureHandler';
 export { TapGestureHandler } from './handlers/TapGestureHandler';
 export { ForceTouchGestureHandler } from './handlers/ForceTouchGestureHandler';
 export { LongPressGestureHandler } from './handlers/LongPressGestureHandler';
@@ -60,10 +49,7 @@ export { PinchGestureHandler } from './handlers/PinchGestureHandler';
 export { RotationGestureHandler } from './handlers/RotationGestureHandler';
 export { FlingGestureHandler } from './handlers/FlingGestureHandler';
 export { default as createNativeWrapper } from './handlers/createNativeWrapper';
-export type {
-  NativeViewGestureHandlerPayload,
-  NativeViewGestureHandlerProps,
-} from './handlers/NativeViewGestureHandler';
+export type { NativeViewGestureHandlerProps } from './handlers/NativeViewGestureHandler';
 export { GestureDetector } from './handlers/gestures/GestureDetector';
 export { GestureObjects as Gesture } from './handlers/gestures/gestureObjects';
 export type { TapGestureType as TapGesture } from './handlers/gestures/tapGesture';
@@ -89,7 +75,7 @@ export type {
   BaseButtonProps,
   RectButtonProps,
   BorderlessButtonProps,
-} from './components/GestureButtons';
+} from './components/GestureButtonsProps';
 export {
   RawButton,
   BaseButton,
